+++ src/api/java/blusunrize/immersiveengineering/api/energy/GeneratorFuel.java
@@ -17,6 +17,7 @@ import blusunrize.immersiveengineering.api.utils.FastEither;
 import blusunrize.immersiveengineering.api.utils.TagUtils;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.ItemStack;
@@ -26,7 +27,10 @@ import net.minecraftforge.registries.RegistryObject;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.Comparator;
 import java.util.List;
+import java.util.SortedMap;
+import java.util.TreeMap;
 import java.util.function.Function;
 
 public class GeneratorFuel extends IESerializableRecipe
@@ -91,5 +95,16 @@ public class GeneratorFuel extends IESerializableRecipe
 			if(fuel.matches(in))
 				return fuel;
 		return null;
+	}
+
+	public static SortedMap<Component, Integer> getManualFuelList(Level level)
+	{
+		SortedMap<Component, Integer> map = new TreeMap<>(
+				Comparator.comparing(Component::getString, Comparator.naturalOrder())
+		);
+		for(GeneratorFuel recipe : RECIPES.getRecipes(level))
+			for(Fluid f : recipe.getFluids())
+				map.put(f.getFluidType().getDescription(), recipe.getBurnTime());
+		return map;
 	}
 }